// front/src/app/swagger/model/categoryResponseDtoPagedResult.ts
/**
 * ExamenTec API
 *
 * Contact: support@examentec.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { CategoryResponseDto } from './categoryResponseDto';


export interface CategoryResponseDtoPagedResult { 
    items?: Array<CategoryResponseDto> | null;
    page?: number;
    pageSize?: number;
    totalCount?: number;
    readonly totalPages?: number;
    readonly hasPreviousPage?: boolean;
    readonly hasNextPage?: boolean;
}